+++ frontend/.env.development
@@ -1,2 +1,2 @@
 REACT_APP_API_URL=http://www.job-profile.com
-REACT_APP_SKIP_AUTH=true
+REACT_APP_SKIP_AUTH=false
