+++ frontend/jsconfig.json
@@ -5,5 +5,5 @@
         "@/*": ["*"]
       }
     },
-    "include": ["src"]
+    "include": ["src", "__tests__"]
   }
