+++ frontend/package.json
@@ -41,5 +41,10 @@
     "postcss": "^8.4.49",
     "prettier": "^3.5.3",
     "tailwindcss": "^3.4.16"
+  },
+  "jest": {
+    "moduleNameMapper": {
+      "react-router-dom": "<rootDir>/node_modules/react-router-dom"
+    }
   }
 }
