+++ frontend/src/services/ai/index.js
@@ -1,13 +1,10 @@
-import createHttpClient from '@/config/HttpClinet';
+import createHttpClient from '@/config/HttpClient';
 
 const createAiService = (hc = createHttpClient('v1')) => ({
-  generate: async userInfo => {
-    const response = await hc.postForm('/ai/generate', { userInfo });
-    if (response.data.token) {
-      localStorage.setItem('token', response.data.token);
-    }
-    return response;
-  },
+    generate: async userInfo => {
+        const response = await hc.post('/ai/generate', { personalExperienceDescription: userInfo });
+        return response;
+    },
 });
 
 export default createAiService;
